+++ core/core.pl
@@ -206,7 +206,10 @@
 :- dynamic('$lgt_hook_goal_expansion_'/2).
 
 
-% multi-threading tags
+% counters
+
+% '$lgt_dynamic_entity_counter_'(Kind, Base, Count)
+:- dynamic('$lgt_dynamic_entity_counter_'/3).
 
 % '$lgt_threaded_tag_counter_'(Tag)
 :- dynamic('$lgt_threaded_tag_counter_'/1).
@@ -1032,7 +1035,7 @@ create_object(Obj, Relations, Directives, Clauses) :-
 
 '$lgt_create_object'(Obj, Relations, Directives, Clauses) :-
 	(	var(Obj) ->
-		'$lgt_gen_entity_identifier'(o, Obj)
+		'$lgt_gen_entity_identifier'(object, Obj)
 	;	true
 	),
 	'$lgt_tr_object_identifier'(Obj, Relations),
@@ -1081,7 +1084,7 @@ create_category(Ctg, Relations, Directives, Clauses) :-
 
 '$lgt_create_category'(Ctg, Relations, Directives, Clauses) :-
 	(	var(Ctg) ->
-		'$lgt_gen_entity_identifier'(c, Ctg)
+		'$lgt_gen_entity_identifier'(category, Ctg)
 	;	true
 	),
 	'$lgt_tr_category_identifier'(Ctg),
@@ -1135,7 +1138,7 @@ create_protocol(Ptc, Relations, Directives) :-
 
 '$lgt_create_protocol'(Ptc, Relations, Directives) :-
 	(	var(Ptc) ->
-		'$lgt_gen_entity_identifier'(p, Ptc)
+		'$lgt_gen_entity_identifier'(protocol, Ptc)
 	;	true
 	),
 	'$lgt_tr_protocol_identifier'(Ptc),
@@ -1152,19 +1155,24 @@ create_protocol(Ptc, Relations, Directives) :-
 
 
 
-% '$lgt_gen_entity_identifier'(+char, -entity_identifier)
+% '$lgt_gen_entity_identifier'(+atom, -entity_identifier)
 %
 % generates a new, unique, entity identifier by appending an integer to a base char
+%
+% note that it's possible to run out of (generated) entity identifiers when using a
+% back-end Prolog compiler with bounded integer support
 
-'$lgt_gen_entity_identifier'(Base, Identifier) :-
+'$lgt_gen_entity_identifier'(Kind, Identifier) :-
+	retract('$lgt_dynamic_entity_counter_'(Kind, Base, Count)),
 	char_code(Base, Code),
 	repeat,
-		'$lgt_next_integer'(1, New),
+		'$lgt_next_integer'(Count, New),
 		number_codes(New, Codes),
 		atom_codes(Identifier, [Code| Codes]),
 	\+ '$lgt_current_protocol_'(Identifier, _, _, _, _),
 	\+ '$lgt_current_object_'(Identifier, _, _, _, _, _, _, _, _, _, _),
 	\+ '$lgt_current_category_'(Identifier, _, _, _, _, _),
+	asserta('$lgt_dynamic_entity_counter_'(Kind, Base, New)),
 	!.
 
 
@@ -19109,6 +19117,17 @@ current_logtalk_flag(Flag, Value) :-
 
 
 
+% '$lgt_initialize_dynamic_entity_counters'
+%
+% counters used when generating identifiers for dynamically created entities
+
+'$lgt_initialize_dynamic_entity_counters' :-
+	assertz('$lgt_dynamic_entity_counter_'(object,   o, 1)),
+	assertz('$lgt_dynamic_entity_counter_'(protocol, p, 1)),
+	assertz('$lgt_dynamic_entity_counter_'(category, c, 1)).
+
+
+
 % '$lgt_load_default_entities'
 %
 % loads all default built-in entities if not already loaded (when embedding
@@ -19282,6 +19301,7 @@ current_logtalk_flag(Flag, Value) :-
 % Logtalk runtime initialization goal
 
 :- initialization((
+	'$lgt_initialize_dynamic_entity_counters',
 	'$lgt_load_default_entities',
 	'$lgt_load_settings_file'(Result, Flag),
 	'$lgt_print_message'(banner, core, banner),